% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_key.R
\name{add_key}
\alias{add_key}
\title{Adds key information to qPCR data}
\usage{
add_key(key_dir, key_name, mydata, all_key)
}
\arguments{
\item{key_dir}{a path to the directory with the keys}

\item{key_name}{the name of the key to be loaded}

\item{mydata}{your qPCR data}

\item{all_key}{a key showing how the 96W plate location relates to the qPCR 384W replicate locations}
}
\value{
Returns the qPCR data with IDs in a column called 'pid'
}
\description{
Adds key information to qPCR data
}
\examples{

}
