% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPO.R
\name{assignPO}
\alias{assignPO}
\title{Assigns pipetting order based on well position}
\usage{
assignPO(data.deltaCT)
}
\arguments{
\item{mydata_deltaCT}{a data frame that contains deltaCT information.}
}
\value{
Returns results of the rank-order analysis
}
\description{
Assigns pipetting order based on well position
}
\examples{
	ranks <- getRanks(mydata_deltaCT)
}
