% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_nas.R
\name{count_nas}
\alias{count_nas}
\title{Counts the number of NAs in a vector}
\usage{
count_nas(to_count)
}
\arguments{
\item{to_count}{vector to count NAs for}
}
\value{
returns number of nas
}
\description{
Counts the number of NAs in a vector
}
\examples{
count_nas(with.pheno.noblue$center)
}
