% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcr_process.R
\name{ddpcr_process}
\alias{ddpcr_process}
\title{Shapes the ddPCR data into a better format for ggplot}
\usage{
ddpcr_process(testdata)
}
\arguments{
\item{testdata}{the original data frame after reading in the csv file}
}
\value{
Returns a data frame that's combined the mt-ND1 and RPPH probe info
}
\description{
Shapes the ddPCR data into a better format for ggplot
}
\examples{
combined.data <- ddpcr_process(testdata)
}
