% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_rug.R
\name{density_rug}
\alias{density_rug}
\title{Function that makes a ggplot density plot with a rug on the bottom}
\usage{
density_rug(to_plot, density_val)
}
\arguments{
\item{to_plot}{Data frame containing whatever you want to plot}

\item{density_val}{A character string with the value you want the density plot for}

\item{col}{The color of the density line}

\item{outline}{If you want the histogram to have an outline}
}
\value{
Returns a ggplot
}
\description{
Function that makes a ggplot density plot with a rug on the bottom
}
\examples{
density_rug(with.pheno.noblue, 'deltaCT.Plateadjust.mean', outline = F)
}
