% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterRPPH.R
\name{filterRPPH}
\alias{filterRPPH}
\title{Filters out samples that have CT.RPPH1 > 5SD away from mean}
\usage{
filterRPPH(data.deltaCT.PO.clean, platename)
}
\arguments{
\item{platename}{the name of the plate that you are working with}

\item{data.deltaCT}{a data frame that contains deltaCT information.}

\item{removed}{a file path that you would like to write removed samples to}
}
\value{
Returns a data frame that has removed samples with deltaCT > 0.5
}
\description{
Filters out samples that have CT.RPPH1 > 5SD away from mean
}
