% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterSD.R
\name{filterSD}
\alias{filterSD}
\title{Filters out samples that have deltaCTs with a standard deviation > 0.5
Filters out the most deviant replicate first, and then if SD still > 0.5, excludes entire sample}
\usage{
filterSD(mydata_withPO, data.SD.means, platename)
}
\arguments{
\item{platename}{the name of the plate that you are working with}

\item{data.deltaCT}{a data frame that contains deltaCT information.}

\item{removed}{a file path that you would like to write removed samples to}
}
\value{
Returns a data frame that has removed samples with deltaCT > 0.5
}
\description{
Filters out samples that have deltaCTs with a standard deviation > 0.5
Filters out the most deviant replicate first, and then if SD still > 0.5, excludes entire sample
}
