% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.outliers.R
\name{find.outliers}
\alias{find.outliers}
\title{Function that will find outliers in a vector (default: > 3 SD from mean), and return their indices}
\usage{
find.outliers(x, sd = 3)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The indices of the outliers
}
\description{
Function that will find outliers in a vector (default: > 3 SD from mean), and return their indices
}
\examples{
find.outliers(mtcars$mpg)
}
