% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatbeta.R
\name{formatbeta}
\alias{formatbeta}
\title{Function that formats betas to look nice}
\usage{
formatbeta(beta)
}
\arguments{
\item{beta}{The value you would like to format}
}
\description{
Function that formats betas to look nice
}
\examples{
formatbeta(beta)
}
