% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatpval.R
\name{formatpval}
\alias{formatpval}
\title{Function that formats betas to look nice}
\usage{
formatpval(pval)
}
\arguments{
\item{pval}{The value you would like to format}
}
\description{
Function that formats betas to look nice
}
\examples{
formatpval(pval)
}
