% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllDups.R
\name{getAllDups}
\alias{getAllDups}
\title{Shapes the ddPCR data into a better format for ggplot}
\usage{
getAllDups(x)
}
\arguments{
\item{x}{the vector you're looking for duplicates in}
}
\value{
Returns indices of all duplicates
}
\description{
Shapes the ddPCR data into a better format for ggplot
}
\examples{
dup_indices <- getAllDups(x)
}
