% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaCT.R
\name{getDeltaCT}
\alias{getDeltaCT}
\title{Reads qPCR data into a dataframe}
\usage{
getDeltaCT(mydata, nuc.probe = "RPPH", mit.probe = "ND1")
}
\arguments{
\item{mydata}{the data frame with the qPCR data}

\item{nuc.probe}{the name of the nuclear probe}

\item{mit.probe}{the name of the mitochondrial probe}
}
\value{
Returns a data frame with the column (deltaCT), which represents the Ct(nuc) - Ct(mito)
}
\description{
Reads qPCR data into a dataframe
}
\examples{
mydata2 <- getDeltaCT(mydata)
}
