% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalMeans.R
\name{getFinalMeans}
\alias{getFinalMeans}
\title{Gets the final means after adjusting for pipetting order}
\usage{
getFinalMeans(data.deltaCT.PO.clean, platename)
}
\arguments{
\item{platename}{the name of the plate that you are working with}

\item{data.deltaCT}{a data frame that contains deltaCT information.}
}
\value{
Returns a data frame that has removed samples with deltaCT > 0.5
}
\description{
Gets the final means after adjusting for pipetting order
}
