% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanSD.R
\name{getMeanSD}
\alias{getMeanSD}
\title{Gets the mean and standard deviation of deltaCT for each sample}
\usage{
getMeanSD(data.deltaCT, removed, platename)
}
\arguments{
\item{data.deltaCT}{a data frame that contains deltaCT information.}

\item{removed}{a file path that you would like to write removed samples to}

\item{platename}{the name of the plate that you are working with}
}
\value{
Returns results of the rank-order analysis
}
\description{
Gets the mean and standard deviation of deltaCT for each sample
}
\examples{
	ranks <- getRanks(mydata_deltaCT)
}
