% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRanks.R
\name{getRanks}
\alias{getRanks}
\title{Performs rank-order analysis on qPCR triplicates.  
Interpretation: significant pipetting order effect = 1st replicate pipetted is enriched for largest (max) deltaCT value. And 3rd replicate pipetted is enriched for smallest (min) deltaCT value}
\usage{
getRanks(mydata_deltaCT)
}
\arguments{
\item{mydata_deltaCT}{a data frame that contains deltaCT information.}
}
\value{
Returns results of the rank-order analysis
}
\description{
Performs rank-order analysis on qPCR triplicates.  
Interpretation: significant pipetting order effect = 1st replicate pipetted is enriched for largest (max) deltaCT value. And 3rd replicate pipetted is enriched for smallest (min) deltaCT value
}
\examples{
	ranks <- getRanks(mydata_deltaCT)
}
