% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_test.R
\name{lm_test}
\alias{lm_test}
\title{Function that will allow for (apply) to be used for linear regressions}
\usage{
lm_test(lm_data, count, covariates = "", correct_for = "")
}
\arguments{
\item{lm_data}{Our dependent variables.  Column = dependent variable, rows = observations (in this case, individuals) (numeric vector)}

\item{count}{Our independent variable (numeric vector)}

\item{covariates}{Covariates associated with each individual. Columns = covariates, rows = observations. (data.frame)}

\item{to_correct}{A string containing additional covariates to correct for.  ex: '+as.factor(covariates$DTHHRDY)+as.factor(covariates$RACE)'}
}
\value{
Coefficients from the regression
}
\description{
This function will do a single linear regression and extract the coefficients from the regression.
}
