% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.R
\name{look}
\alias{look}
\title{Function that gives you the first five columns and first five rows of a dataframe}
\usage{
look(df)
}
\arguments{
\item{df}{The data frame you wanna look at}
}
\description{
Function that gives you the first five columns and first five rows of a dataframe
}
\examples{
look(df)
}
