% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look.resid.R
\name{look.resid}
\alias{look.resid}
\title{Looks at the residuals vs. predicted values for a model}
\usage{
look.resid(data, model)
}
\arguments{
\item{data}{the data you are trying to do models on}

\item{model}{the model that you would like to look at}
}
\value{
returns number of nas
}
\description{
Looks at the residuals vs. predicted values for a model
}
\examples{
look.resid(no surgery, model = lm(mito_percents ~ sex, data = no_surgery))
}
