% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looknd.R
\name{looknd}
\alias{looknd}
\title{Function that will look at the ND1 copies in ddPCR data}
\usage{
looknd(good_data)
}
\arguments{
\item{good_data}{data with enough droplets.}
}
\value{
A ggplot showing the amount of mt-nd1 in each sample
}
\description{
Function that will look at the ND1 copies in ddPCR data
}
\examples{
looknd(chemicaldata)
}
