% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookrpph.R
\name{lookrpph}
\alias{lookrpph}
\title{Function that will look at the RPPH1 copies in ddPCR data}
\usage{
lookrpph(good_data)
}
\arguments{
\item{good_data}{data with enough droplets.}
}
\value{
A ggplot showing the amount of rpph1 in each sample
}
\description{
Function that will look at the RPPH1 copies in ddPCR data
}
\examples{
lookrpph(chemicaldata)
}
