% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.forestplot.R
\name{make.forestplot}
\alias{make.forestplot}
\title{Function for pretty forestplots, adapted from Ryan}
\usage{
make.forestplot(
  model.slot = 4,
  numfits = 4,
  model_names,
  fit1,
  fit2,
  fit3,
  fit4,
  inc = 0.1
)
}
\arguments{
\item{model.slot}{the position of the thing you're comparing to within the model summary results}

\item{numfits}{The number of models you're looking at}

\item{model_names}{The names for your models in a vector of characters}

\item{fit1}{The first model}

\item{fit2}{The second model (and so on and so forth.  you need to put in four fits.)}

\item{inc}{The increments for the xticks}
}
\value{
A nice forestplot
}
\description{
Function for pretty forestplots, adapted from Ryan
}
\examples{
model_names <- c("coxph(survival) ~ mtDNA-CN", "Model 1 + Age, Sex, Race", "Model 2 + Hypertension, Diabetes, Smoking", 'Model 3 + BMI, SBP, Troponin, AF history')
make.forestplot(model_names = model_names, fit1 = fit1, fit2 = fit2, fit3 = fit3, fit4 = fit4)
}
