% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.quintiles.R
\name{make.quintiles}
\alias{make.quintiles}
\title{Function to make quintiles from a column within a dataframe --> taken from fashaR}
\usage{
make.quintiles(test)
}
\arguments{
\item{test}{data that needs to be divided into quintiles}
}
\value{
output Dataframe with qunitiles (categories and by number, 5 = highest quintile)
}
\description{
Function to make quintiles from a column within a dataframe --> taken from fashaR
}
\examples{
test2=makequintiles(test$resid.mtDNA)
}
