% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quintiles.R
\name{makequintiles}
\alias{makequintiles}
\title{Function to make quintiles from a column within a dataframe --> taken from fashaR}
\usage{
makequintiles(test, x)
}
\arguments{
\item{test}{Dataframe that contains vector to divide into quintiles}

\item{x}{Name of column that will be divided into quintiles}
}
\value{
output Dataframe with additional columns named quintiles and num.quintiles
}
\description{
Modified function form of the function grs.plot from package gtx
that allows coloring by allele frequency.
}
\examples{
test2=makequintiles(test, resid.mtDNA)
}
