% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit.outliers.R
\name{omit.outliers}
\alias{omit.outliers}
\title{Function that will find outliers in a vector (default: > 3 SD from mean), set them to be NAs, and return the vector}
\usage{
omit.outliers(x, sd = 3)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The same vector, with outliers changed to be NAs
}
\description{
Function that will find outliers in a vector (default: > 3 SD from mean), set them to be NAs, and return the vector
}
\examples{
omit.outliers(mtcars$mpg)
}
