% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{Function for plotting cor values onto a correlation matrix}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{lm_object}{The results of an lm that you want a p-value QQ-plot for}
}
\value{
A p-value QQ-plot
}
\description{
Function for plotting cor values onto a correlation matrix
}
\examples{
pairs(~TRISCHD + covariates, data = df, na.action = na.omit, upper.panel=panel.cor)
}
