% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plots xvar against yvar, and title is the P-value}
\usage{
plot_lm(df, xvar, yvar)
}
\arguments{
\item{df}{data frame for ggplotting}

\item{character}{of yvariable name}
}
\value{
a ggplot object
}
\description{
Plots xvar against yvar, and title is the P-value
}
\examples{
plot_lm(df, 'xvariable', 'yvariable')
}
