% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qpcr_data.R
\name{read_qpcr_data}
\alias{read_qpcr_data}
\title{Reads qPCR data into a dataframe}
\usage{
read_qpcr_data(data_dir, filename, skiplines = 45)
}
\arguments{
\item{data_dir}{a path to the data directory where your files are stored}

\item{filename}{the name of the file to be loaded}
}
\value{
Returns a dataframe with all the qPCR data in it.
}
\description{
Reads qPCR data into a dataframe
}
\examples{
df <- read_qpcr_data('/Users/arkinglab/Documents/', 'qPCR_data.xls') 
}
