% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrRight.R
\name{substrRight}
\alias{substrRight}
\title{Takes the last n characters of a string.  Taken from: https://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r}
\usage{
substrRight(x, n)
}
\description{
Takes the last n characters of a string.  Taken from: https://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r
}
\examples{
substrRight('character', 3)
}
