% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPOadjust.R
\name{addPOadjust}
\alias{addPOadjust}
\title{Adjusts for pipetting order
	data.deltaCT.PO.clean <- getRanks(mydata_deltaCT)}
\usage{
addPOadjust(data.deltaCT.PO.clean)
}
\description{
Adjusts for pipetting order
	data.deltaCT.PO.clean <- getRanks(mydata_deltaCT)
}
