% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeanSD.R
\name{getMeanSD}
\alias{getMeanSD}
\title{Performs rank-order analysis on qPCR triplicates.  
Interpretation: significant pipetting order effect = 1st replicate pipetted is enriched for largest (max) deltaCT value. And 3rd replicate pipetted is enriched for smallest (min) deltaCT value}
\usage{
getMeanSD(data.deltaCT, removed, platename)
}
\arguments{
\item{data.deltaCT}{a data frame that contains deltaCT information.}

\item{removed}{a file path that you would like to write removed samples to}

\item{platename}{the name of the plate that you are working with}
}
\value{
Returns results of the rank-order analysis
}
\description{
Performs rank-order analysis on qPCR triplicates.  
Interpretation: significant pipetting order effect = 1st replicate pipetted is enriched for largest (max) deltaCT value. And 3rd replicate pipetted is enriched for smallest (min) deltaCT value
}
\examples{
	ranks <- getRanks(mydata_deltaCT)
}
