% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give.n.R
\name{give.n}
\alias{give.n}
\title{Allows boxplots to output the number of observations for each factor.}
\usage{
give.n(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Allows boxplots to output the number of observations for each factor.
}
\examples{
p <- ggplot(all2, aes(x=Group2, y=corr_mito_D)) + geom_boxplot() + stat_summary(fun.data = give.n, geom = "text")
}
