% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_test.R
\name{lm_test}
\alias{lm_test}
\title{Function that will allow for (apply) to be used for linear regressions}
\usage{
lm_test(lm_data, count, covariates = "", correct_for = "")
}
\arguments{
\item{lm_data}{Our dependent variables.  Column = dependent variable, rows = observations (in this case, individuals) (numeric vector)}

\item{count}{Our independent variable (numeric vector)}

\item{covariates}{Covariates associated with each individual. Columns = covariates, rows = observations. (data.frame)}

\item{to_correct}{A string containing additional covariates to correct for.  ex: '+as.factor(covariates$DTHHRDY)+as.factor(covariates$RACE)'}
}
\value{
Coefficients from the regression
}
\description{
This function will do a single linear regression and extract the coefficients from the regression.
}
\examples{
for_test <- blood_full[,1:20702] # the transcripts you're testing as your dependent variables
covariates <- blood_full[,20703:21001] # the covariates
MT_count <- as.numeric(blood_full[,21002]) # the MT_count
lm_results <- apply(for_test, 2, lm_test, count = MT_count, covariates = covariates, correct_for = '+as.factor(covariates$DTHHRDY)+as.factor(covariates$RACE)')
}
