% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.cor.R
\name{panel.cor}
\alias{panel.cor}
\title{Function for plotting cor values onto a correlation matrix}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\value{
text for the pairwise correlation plot
}
\description{
Function for plotting cor values onto a correlation matrix
}
\examples{
pairs(~PC1 + PC2, data = df, lower.panel=twolines, diag.panel=mydiag.panel, upper.panel=panel.cor, label.pos=0.5, main="Correlation matrix of PCs and covariates", labels = c('PC1', 'PC2'))
}
