% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGroup.R
\name{sampleGroup}
\alias{sampleGroup}
\title{Function for sampling, taken from stackoverflow}
\usage{
sampleGroup(df, size)
}
\arguments{
\item{df}{the df to sample from}

\item{size}{number of samples}
}
\description{
Function for sampling, taken from stackoverflow
}
\examples{
Will sample 2 observations per species in the iris dataframe
iris.sample<-ddply(iris,.(Species),function(df) sampleGroup(df,2))
}
